Create Creating Basic Web Interface
I have added CSS and JS in the same file. Contrary to some developers , i find it better to keep everything on the same page. You can see from the code that the code is very lucid and only a very basic knowledge is required to understand this. In my opinion the simpler the better